+++ lib/hive/customergroupgroupspecialprice_hive.dart
@@ -14,7 +14,7 @@ class CustomerGroupGroupSpecialPrice extends HiveObject {
   late String groupCode;
 
   @HiveField(3)
-  late double disc;
+  late int disc;
 
   @HiveField(4)
   late String notes;
