+++ lib/hive/customerpropbrandspecialprice_hive.dart
@@ -14,7 +14,7 @@ class CustomerPropBrandSpecialPrice extends HiveObject {
   late String brandCode;
 
   @HiveField(3)
-  late double disc;
+  late int disc;
 
   @HiveField(4)
   late String notes;
